Move new file to utils folder
