+++ web/js/utils/generalUtils.js
